+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Ability to fill a struct from another one (`fill_from` method)
 - Ability to extract data from a struct as a dictionary (`to_dict` method)
+- Ability to clear the buffer a struct points to (`clear_buffer` method)
 
 ## [1.2.1] - 2024-11-30